requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.4

# HTTP client
httpx==0.27.0

# Config & validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Scheduling
apscheduler==3.10.4

# LLM
google-generativeai==0.7.2

# Logging
structlog==24.1.0

# Utils
python-dotenv==1.0.1
PyYAML==6.0.1
pybloom-live==4.0.0
tenacity==8.3.0
